+++ src/LinkDotNet.LinqSIMDExtensions/Min.cs
@@ -42,6 +42,11 @@ public static partial class LinqSIMDExtensions
     public static T Min<T>(this Span<T> span)
         where T : unmanaged, IMinMaxValue<T>, INumber<T>
     {
+        if (span.IsEmpty)
+        {
+            throw new InvalidOperationException("Sequence contains no elements");
+        }
+
         var spanAsVectors = MemoryMarshal.Cast<T, Vector<T>>(span);
         var minVector = VectorHelper.CreateWithValue(T.MaxValue);